CC3XX: Add config switches for Chacha20/Poly1305 and CCM
Add configuration switches for Chacha20/Poly1305 and CCM in the CC3XX
driver. For the time being they're still hooked to the mbed TLS
config file but will have to adapted to the config strategy when
that is finalised (and likely to move to PSA_WANT_* mechanism).
Move the GCM ones from the internal layer to the interface.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: Ib8620d8cbf8ddb3981bb975ffc93c7b62f8a5aa7