Crypto: Enable MM-IOVEC
The MM-IOVEC is a feature of FF-M v1.1 to allow Secure Partitions
to access client buffers directly.
With this feature, Secure Partitions do not need transient buffers
to store client buffers so that can save RAM size.
And there is also no memory copy so the performance should be better
as well.

This patch also re-organizes the codes for better reading.

Change-Id: I488d046a53dc6a122a3b2b59adc360448a8e4bdb
Signed-off-by: Kevin Peng <kevin.peng@arm.com>