SFN: Support initialize SFN partitions by priority
IPC partitions get initialized by the priority, the SFN partitions
also has this requirement in some cases. For example, secure test
partition is inited before basic partitions (Crypto, ITS).This
causes an issue that the service timing of first call to request
a partition is very large because it includes the partition
initialization process. To get accurate performance data, we should
deprecate the first call data.

In order to simplify the performance test, the basic partition
initialization process had better be in front of secure test partition.

Signed-off-by: Jianliang Shen <jianliang.shen@arm.com>
Change-Id: I3ce03e971a80826be18ad945a1bd1f96e5201cae