Build: Add option CONFIG_TFM_ENABLE_FPU
Add CONFIG_TFM_ENABLE_FPU for enabling the FPU coprocessors when either
or both of the SPE and the NSPE needs to operate the FPU. On an Armv8-M
platform with TrustZone, this option also enables Non-Secure access to
the FPU.

The motivation of this new option is that in the TrustZone architecture,
both the SPE and NSPE must enable the FPU consistently. If one of the
processing environments has the FPU coprocessors enabled and the other
has not, then a context switch (e.g. an interrupt) that happens during
a cross-environment call will trigger a NOCP (No coprocessor) usage
fault, if CONTROL.FPCA==1 (floating point context active). This fault
happens because CONTROL.FPCA is not banked, and the system see a
contradiction if the current side does not have the FPU enabled but
the floating point context (FPCA) has been set as active by the other
side.

CONFIG_TFM_FP=hard, which enables floating point instructions and the
hard float ABI, depends on CONFIG_TFM_ENABLE_FPU=ON as the FPU is used.
Therefore a dependency check has been added.

For a user of TF-M, setting CONFIG_TFM_ENABLE_FPU=ON while leaving
CONFIG_TFM_FP=soft (default) enables the FPU coprocessors but keeps
the TF-M SPE to be without any floating point operations. The user's
Non-Secure application is allowed to operate the FPU as needed.

Note: If the SPE operates the FPU (CONFIG_TFM_FP=hard), the compiler
support for patching the CVE-2021-35465 (VLLDM instruction security)
vulnerability is required for FPU context protection. If the NSPE is
the only side that operates the FPU, then the FPU context protection
can be skipped as the SPE does not touch the floating point registers
anyway.

Change-Id: I20c80f3d86087585bc569a1bfa183c7c22220c09
Signed-off-by: Lingkai Dong <lingkai.dong@arm.com>