Tools: Pass build configurations through header file
This patch adds a new configuration header file argument for
the manifest tool.

It is used to pass build configurations including:
- isolation level
- backend
- Configurations switches of Secure Partitions

The config file is c header with configurations defined as Macros.
The config file could be generated/maintained by all means.
In the TF-M build system, CMake generates it at build time.

Multiple config files are supported.

With the new config header, the "conditional" attributes should
not contain the "@" symbols.

And configure_file() is not required to get the values of
"conditional" attributes.

Note that configure_file() may be still performed to eliminate
CMake varibles in "manifest" attributes.

Change-Id: I9b1e4a68652c67548a079428e494e313383ec845
Signed-off-by: Kevin Peng <kevin.peng@arm.com>