SPM: Deprecate customized '*_mem*' API
The 'tfm_mem*' APIs in tfm_memory_utils.h was designed for indicating
potential specific mem* implementation in secure firmware. As now the
general mem* API in string.h is callable in secure firmware, deprecate
the legacy implementation as it is unlikely to have an update.

Also limit the 'spm_mem*' API usage in platform sources, because
platform sources are shared between targets instead of being specific
for SPM.

Signed-off-by: Ken Liu <Ken.Liu@arm.com>
Change-Id: Iea731fe14410bd7c319a378474da7b5282a28496