Tool: Detect circular dependency in Secure Partitions
Circular dependencies between Secure Partitions are forbidden in FF-M.

For example, Partition A depends on service of Partition B,
while Partition B depends on Service A1 in Partition A.
That is a circular in partition dependency which is invalid.

This patch adds detection on this kind of circular when processing
partition manifests.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I7d3e7b5636c9463e96686e57019af0fb85181d1f