corstone1000: remove two partition configuration
Previously to run tf-m test, a larger partition was created
which allowed all default test binaries to be included.
The patch revert the change because any partition might
not be enough to hold all test binaries in the future.
So its better to run few test at a time instead of creating
a larger partition.

Signed-off-by: Satish Kumar <satish.kumar01@arm.com>
Change-Id: I223fe45f2de014dbcadc6ac12c321c524701116a