platform: Fix exception handlers not overiding default implementations
Fix regression introduced by: 2e4348963610787aaa0e9cf14fe45af697b6c464.
The commit moved the exception handlers from the tfm_s library to the
platform_s library.
This causes the fault handlers to be discarded since they are not used
and they are not overriding the startup files' weak handlers.
Move the fault handlers to the tfm_s library to fix this problem.

Change-Id: Icea99165fa0867fae1bcdf5589f569435d79da17
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>