SPM: Update stack-switch checking when cross call
When Trustzone NS Agent calls SPM API via cross-stack call ABI, the
stack switching is unnecessary as the agent shares the stack with
SPM. The stack-switch checking was indicated by an argument of the
ABI handler, which needs the caller to tell if a stack-switching
is needed. This mechanism occupies one more argument of the handler.
Actually, the ABI handler can check if the current working stack
belongs to the agent (SPM) to decide a switching is necessary.

To do this, several places are updated:

- Deprecate the member 'allocated' in context control because it has
  less significance. Involve a new member 'sp_start' to indicate the
  stack usage start.
- Add stack-switching checking logic in the ABI handler. And remove
  the original switching indicator.

Also provides an architecture-specific MACRO for creating the spm
context control global variable claiming, as the creating contains
architecture-specific settings.

Signed-off-by: Ken Liu <Ken.Liu@arm.com>
Change-Id: I7a12be36ff9017121e655d2aa328145612d00629