SPM: Add additional check in psa_wait()
In psa_wait(), it is considered as programmer's error if signal_mask
does not include any assigned signals. Before checking this, it is
essential to check if the calling partitions has any assigned signals
or not. For eg idle partition which has no assigned signals calls a
dummy psa_wait() which results in a psa_panic().

Signed-off-by: Ruchika Gupta <ruchika.gupta_1@nxp.com>
Change-Id: I6e5431302970bb3b0be801569172cfc7002992a2