SPM: Clear FP context when initialization ends
Hardware FP may be used during SPM initialization when it is enabled.
The existing FP context clearing is not late enough so that the SPM
execution still can bring FP context after the clearing.

Clear the FP context when SPM initialization ends when hardware FP
is enabled.

The FP status generated is cleared by the status bit in EXC_RETURN.
Hence the dedicated clear FP status function is not used anymore.

Signed-off-by: Ken Liu <Ken.Liu@arm.com>
Change-Id: Ief6855f18d27b26818e04206dadc406ffa07c753