platform: nordic_nrf: Add IOCTL non-secure API source file to install
Add the non-secure API IOCTL functions for the nordic platform to the
set of source files exported in the install folder.
In the case where this is built by an external build system instead of
the platform_ns library then this source file needs to be included in
the non-secure application and its build system.

Change-Id: Icd0312bdc3e583f5eb32cde589e2bc3c9a67ffdc
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>