platform: lairdconnectivity: Add IOCTL non-secure API source file to install
Add the non-secure API IOCTL functions for the Laird Connectivity
platform to the set of source files exported in the install folder.
In the case where this is built by an external build system instead of
the platform_ns library then this source file needs to be included in
the non-secure application and its build system.

Change-Id: I53dfdf48446270f2079d0f710125e9a68db577f7
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>