Attest: Introduce PSA 2.0.0 attestation profile
The attestation profile is evolving.
The original implementation is based on this version:
 - https://www.ietf.org/archive/id/draft-tschofenig-rats-psa-token-05.html
 - Profile name in the config: PSA_IOT_1

An updated version is available:
 - https://www.ietf.org/archive/id/draft-tschofenig-rats-psa-token-09.html
 - Profile name in the config: PSA_2_0_0

This change (and some follow-up) addresses the
differences between these and make the attestation
profile configurable. Currently, both profile is
supported. But in the long-term, the PSA_IOT_1 is
planned to be deprecated and then later removed.

Signed-off-by: Tamas Ban <tamas.ban@arm.com>
Change-Id: Iebe584f22e05fa60a324e0a13126440cb6f936d5