platform: nordic_nrf: Make psa_call/svc unpriv
Add the psa_call_pack and psa_interface_svc to
the unprivilleged part of the image. When
isolation level >1 is used the PSA application
RoT partitions (such as PS) run in unprivilleged
mode and they need to be able to access these
functions when accessing any other RoT services
(such as ITS).

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: Id82e8fadd1822930162b7bb8b1f434891c5f20d2