Fix: Attest HAL default implementation
The 'tfm_plat_otp_read' always copies the smaller
amounts of bytes to the buffer, comparing the size
of the destination buffer and the allocated space in
the OTP for the provisioned data item.

In the attest HAL default implementation might
a wrong value of the actually copied data is returned
to the caller. It depends on the relation between
the size of the output buffer and the allocated
space in the OTP:

- If output buffer is smaller than allocated
  space in OTP, then only the size of the output
  buffer data is copied, there is no overflow,
  but it wrongly indicates that the size of the
  allocated space in the OTP is copied.
- If content is a string then the zero termination
  character might missing from the end of the buffer
  and tfm_strlen wrongly indicates the length of the
  string. It might over search the bounds of the output
  buffer, because tfm_strlen is limited by the size
  of the allocated space in the OTP rather than the
  size of the output buffer.

These faults can lead that random data is added to the
attestation token.

Signed-off-by: Tamas Ban <tamas.ban@arm.com>
Change-Id: I5c06b5ad9902307c81c8a72345a5e7e86cb05b4a