corstone1000: support for UEFI FMP image Information
The commit provides the support for UEFI FMP (Firmware Management
Protocol) SET and GET Image info APIs.

The APIs to SET and GET image info is implemented. In current design,
SET is called by secure encalve and GET is called by the host.

FMP image information is initialized on every boot and retained
in SRAM. The updatable values of the FMP are stored in private
metadata section of the flash.

Change-Id: Iaf0b4a13a9c24f05e4a32509e61a8b96ee8e9e4b
Signed-off-by: Satish Kumar <satish.kumar01@arm.com>