SPM: Fix macro checks for clearing floating point context
The SPM only needs to clear FP context if it is compiled to use hard
FP, i.e. `CONFIG_TFM_FP >= 1`. If `CONFIG_TFM_ENABLE_FPU` is defined
while `CONFIG_TFM_FP == 0`, the SPM itself does not use the FPU but only
enables CP10 and CP11 and makes them accessible to the NSPE. In this
case, clearing of FP context is not needed because the SPM never uses
the FPU, and the code that does the clearing fails to compile because
instructions and registers for floating point are not allowed by the
compiler when not compiling with hard FP.

Change-Id: Ib659119aedb8eb5a0b2543686bec40677091a4af
Signed-off-by: Lingkai Dong <lingkai.dong@arm.com>