Build: Support for configurable stack and heap size
This patch adds support for configurable stack and heap size
of Secure Partitions.

The build system checks if the "stack_size" and "heap_size"
attributes are valid numbers or configurations.
If they are configurations, add them into the config header
file for the manifest tool.

Then the manifest tool checks if they are configrations as well.
If they are, it replaces them with the values in the config
header.

For Secure Partitions, to enable configurable stack or heap size,
they need to:
1. Add a configuration item for stack or heap size with default value
2. In the manifests, replace the stack or heap size value with the
   added configuration.

Change-Id: I99a9b5508dc044ea8137a3cd73a86d796a51c518
Signed-off-by: Kevin Peng <kevin.peng@arm.com>