Platform: Call 'SystemInit' before 'main' in particular platforms
Some platforms changed the default CMSIS Pack codes, so that calling
into 'SystemInit' for NSPE startup is skipped. This leads to a problem
that an ealier initialization before 'main', which is required by some
features (FPU e.g.), cannont be correctly performed.
This patch adds back the missing call to 'SystemInit' for these
platforms.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I012597b29af38483deb7704c9986c3c109d1654e