corstone1000: allocate static space for the virtqueues in openamp
The change removes the dynamic allocation of virtqueues
and uses the static allocation. The dynamic allocation from
stdlibc identified to have allocated memory from the wrong
ram region. Further investigation is needed but for now
dynamic allocation is removed.

The problem came to surface when below commit was pushed
which moved the location of HEAP in the loader file:

  commit e152498ccda0e25544f5429a8f53b9bbc94733ec
  Platform: Remove unused heap

    - Set mbedtls_print to tfm_sp_log_printf which does not need heap.
    - After that in case of no debug information, heap is not needed.
      Remove heap in link scripts.
    - In case of with debug information, heap is used by GNU c lib
      functions like assert, vprintf and so on. Reserve 0x200 bytes
      heap for that.

Change-Id: I5c620ba0424551ab379bcebd999b117dc565ccb9
Signed-off-by: Satish Kumar <satish.kumar01@arm.com>