Attest: Move label definitions to public header
These defines are needed internally for encoding the initial attestation
token, but are also used for testing when the token is decoded and
verified (can be used by a verifier entity for the same reason).
This patch moves these macro definitions to a public header file
of the partition. The same defines are used in the Delegated
Attestation partition as well.

Change-Id: I23c331902d810fceaf76602dc90c2ae7ae7e7d84
Signed-off-by: David Vincze <david.vincze@arm.com>