CC3XX: Fix Poly driver memory corruption issue
The memcopy function copied all the accumulator
data to the tag in PolyFinish function. The
accumulator data of the PolyPkaContext_t is
20 bytes but the tag buffer is only 16 bytes.
This caused a memory corruption issue.
This patch makes sure that the tag data
are filled by the accumulator base on the
size of the tag buffer. Also, makes sure that
the tag buffer is expected to be exactly as 16
bytes.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: Ieb2fb61b1fbc9e8aecd3db9784485093bc9702f4