Attest: Use measurement size from header
Changes the attestation partition to use MEASUREMENT_VALUE_MAX_SIZE from
the Measured Boot header as the measurement buffer size. The previously
used MEASUREMENT_VALUE_SIZE is a build variable from the Measured Boot
partition, so depends on some CMake magic to be present in the
attestation partition. Removing it simplifies the build at the cost of
a small amount of extra stack when Measured Boot is enabled and the
actual measurement size is less than the maximum.

Signed-off-by: Jamie Fox <jamie.fox@arm.com>
Change-Id: I70ead7ee3840cb753c6a4b32e75d92a3088b93eb