Platform: Check that ProgramData programmed all the data
ProgramData returns the number of data items programmed or a status
error code. Status error codes may be 0 to indicate success or
negative to indicate failure.

Presumably the API may return the number of bytes written so that
large writes can be retried on partial write failure without having to
re-write all of the data.

In this patch we add check for a partial write.

Change-Id: I434c7ac358237c72ea3c8f542e7c45783566b4f1
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>