platform: otp: Simplify otp write and improve doc
We have a check like this:

if ((buffer[idx] | in[idx]) != in[idx]) {
  return TFM_PLAT_ERR_INVALID_INPUT;
}

where 'buffer' is the old OTP value and 'in' is the value we are
attempting to write.

This check ensures that we don't write any bit from 1 to 0.

Since we have this check it is safe to directly write the 'in' buffer
to OTP. It is not necessary to write the 'in' array to the buffer.

Also, misc. doc fixes and comment improvements.

Change-Id: I9b226cd11d770c7532d87ac748ded53860c1e926
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>