Tool: Remove hard dependency on C headers
The manifest parser is expected to run independently but the hard
dependency in the parser makes this not applicable.

For the static handle generating factors, use hard code at present
until this generating scheme is changed to support more static handle
numbers larger than 32 in a flexible way, which is unlikely to happen
in a short time.

As explained in the tooling source, use the same factor names both in
C and the tooling source, because:

* The scheme and factors are totally the same for C and tooling but
  written in different languages.
* When the number increased to larger than 32, the scheme itself needs
  an update besides factors - it is not just a 'change the factor'
  thing, hence making only the factors configurable can not work well.
* The name is a reminder - when developer search for the variables,
  they would spot all the places and update them together.

Signed-off-by: Ken Liu <Ken.Liu@arm.com>
Change-Id: Ic0f7131c12c77c9f8da00a978110a7887db59c83