platform: nrf: Lock network core as non-secure
The network core bus access is configured via the SPU external domain
register. By default this is set to Non-Secure but it can be set to Secure
by the bootloader to perform a DFU of the network core. So it gets set
to non-secure by TF-M, in case the bootloader didn't set it back to
non-secure again.
As the DFU happens before TF-M is booted, the SPU EXTDOMAIN register can
also be locked, so it can't be changed until the next reset. Therefore
it's ensured that it can't be switched back to secure by accident within
TF-M.

Change-Id: I8c278901c1deb150b1cb531fd4faff358f8b5568
Author: Markus Swarowsky <markus.swarowsky@nordicsemi.no>
Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>