Build: Allow TFM_DUMMY_PROVISIONING for non-PLATFORM_DEFAULT_OTP
Currently you get an invalid configuration when TFM_DUMMY_PROVISIONING
is enabled and PLATFORM_DEFAULT_OTP_WRITEABLE is disabled.

But we wish to use TFM_DUMMY_PROVISIONING when PLATFORM_DEFAULT_OTP is
disabled, and in this configuration it doesn't make sense to have
PLATFORM_DEFAULT_OTP_WRITEABLE enabled.

We could alternatively introduce a TFM_NRF_DUMMY_PROVISIONING, but we
would prefer to not have two configs mean the same thing.

Change-Id: Ib9d46cd0de58942a91b787eb4ba9b99b04354f26
Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>