RSS: Enable measured_boot_read() from AP
To be able to execute the measured boot regression test
suite on the AP core, the measured_boot_read() call
must be enabled on the AP-RSS interface.

There is no security implication if AP core can
read the measurements. In normal operation it is not
done, just in the regression test suite.

Signed-off-by: Tamas Ban <tamas.ban@arm.com>
Change-Id: Icd86a33ba41279dda25706180e29e8846e7ec663