SPM: Improve SPM_THREAD_CONTEXT logic
When ns_agent_tz is used, SPM shares a stack with it. When
ns_agent_mailbox is used, SPM needs its own dedicated stack.
The correct logic, then, is to create the dedicated SPM stack only if
CONFIG_TFM_USE_TRUSTZONE is not set, and to only set SPM_THREAD_CONTEXT
to the ns_agent_tz context rather than that of either ns_agent.

Change-Id: Ibac9283f7e5cf7b746804754f6e1aaf8199d9f6c
Signed-off-by: Chris Brand <chris.brand@cypress.com>