QCBOR: restore q_useful_buf.h file
The q_useful_buf.h is a wrapper for the UsefulBuf.h header
in the QCBOR library to apply TF-M coding style. This
patch restores this header as it is still needed for components
that only links to qcbor, but not to the t_cose library
(which still includes a wrapper).
The header was removed by this commit: 8836b222

Change-Id: Ib91d454d594b988a1c7d506cdf6e6a33c2a90de8
Signed-off-by: David Vincze <david.vincze@arm.com>