QCBOR: Disable QCBOR float usage
Disable the QCBOR float usage.
In toolchains without fenv support this will cause a compilation error
because of the FE exception types e.g. FE_DIVBYZERO are not defined
in the fenv.h headers.
Float types are not currently used by the dependencies of QCBOR so this
should be fine and generates a small code decrease.

Change-Id: I537b283188cfab2cf4cc22890da7b9e479fd71fc
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>