Build: Fix compilation warning on array subscript of stack seal region
The stack seal region is defined as an 8 byte region in the linker
scripts, then declared as an extern uint32_t i.e 4 bytes value.

The compiler generates a warning when this is cast to a uint64_t *
in order to write the 8 byte stack seal value.

Correctly declare the stack seal region as an 8 byte region.

The CMSIS stack seal function accepts an uint32_t pointer then
itself cast it back to a uint64_t pointer.

Change-Id: I0e0a490cb3dcba83b7c506457b2e001dcc48febf
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>