SPM: Fix incorrect stack seal of PSP stack
Fix the writing of the PSP stack seal writing the seal value to the
incorrect memory location.
The calculation is a copy paste mistake of the TFM secure stack seal
calculation, which uses - 2 to calculate from the limit instead of the
base, and [-1] to reserve space for IOVEC struct (struct iovec_args_t).

Fix spelling in variable name.

Change-Id: Ia491bd0e3b7708fa379d6e87bbcba1392579c8e2
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>