Config: Disable Doorbell API for SFN backend
Doorbell APIs are not supported in SFN backend. Though CONFIG_TFM_DOORBELL_API
is redefined to 0 in config_spm.h when SFN backend is selected, there
will be a warning for the redefinition. To avoid this warning message,
better to disable Doorbell APIs in project config header files if SFN is
the default backend.

This patch does the following things:
1. Change CONFIG_TFM_DOORBELL_API to 0 in all project config header files
that use SFN backend.
2. Add invalid config check in config_spm.h to make sure that doorbell
APIs are disabled when SFN backend is used.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I46978643185fa9756f657cc8d3c780811892cab0