Kconfig Tool: Support input ENV. and config files loading
The Kconfig tool now supports loading config files and merging them
into one. The merging is basically done by the later loaded config files
overriding the first loaded config files with dependencies respected.

With this feature, the Kconfig system will be aware of the configs such
as platform settings and profile settings rather than having default
values for all configurations.

The environment variables are for the Kconfig files which refer to env.
variables, for example the PLATFORM_PATH.
Now the PLATFORM_PATH is no longer passed as one of the arguments, it's
done via the --env argument now. Of course, any other env. variables can
be appended at will.

Change-Id: Ib39e13b29def382a9546cd7680c62a4f4d321f99
Signed-off-by: Kevin Peng <kevin.peng@arm.com>