CC3XX: Fix macro in PSA sign_message
The macro PSA_ALG_ECDSA_IS_DETERMINISTIC just checks
(if alg & 0x100 != 0). This is not correct since it
does not check if the algorithm is ECDSA in the
first one.
The macro which replaced it is the
PSA_ALG_IS_DETERMINISTIC_ECDSA which firstly checks
if the algorithm is ECDSA and then checks if it is
deterministic.

This affects the function when call with RSA PSS
as an algorithm since it has the value:
PSA_ALG_RSA_PSS_BASE 0x06000300

Which means that RSA PSS algorithm will be
interpreted as ECDSA determinstic and it will
return PSA_ERROR_NOT_SUPPORTED.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
Change-Id: I750415c4d82009f10a0c2997c1ee5abf8e1696b7