Build: Change some CACHE variables to normal ones
Some platform configurations are fixed to platforms and integrations
but the types are CACHE.
CACHE variables are usually considered to be user changable.

This patch changes those variables to normal variables.
Since the platform configs are "included" by the root CMakeLists.txt,
the scope of them are root CMakeLists.txt as well and then can be
seen by other CMake files.

Change-Id: I6ec4eb80363cdb6c62b7e1c8391a8c9265f6e477
Signed-off-by: Kevin Peng <kevin.peng@arm.com>