platform: nordic_nrf: Align nordic_nrf_s.ld with tfm_common_s.ld
Update the nordic_nrf_s.ld to align it with tfm_common_s.ld.
The only differences between these files should be the nordic specific
changes.
The Veneers are placed in a different location because of the Nordic
specific SPU, instead of the ARM SAU.
The NULL pointer detection feature.
The regions Secondary and Storage are defined according to the Nordic
platform configurations.

Compare:
vim -d \
./platform/ext/common/gcc/tfm_common_s.ld \
./platform/ext/target/nordic_nrf/common/core/gcc/nordic_nrf_s.ld

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>