platform: nordic_nrf: Use the common linker files
Use the common linker scripts for the nordic platform.
The nordic linker file for secure has become out of sync with the
changes made to the linking and enabling of HEAP, resulting in
compilation errors when using isolation level 2.
This fixes this issue by using the common linker file instead.

Change-Id: I0fa68b6d8560922cf84b6b51e1b9f45125c95f55
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>