Kconfig: Add the ability to convert cache vars to .config
This patch adds a function to goes through the CMake cache variables
and convert them to .config format.
With this function, the Kconfig system can use it to get the
config options set by platform in config.cmake.

The .config files will be loaded by the tfm_kconfig.py to be
visible to users.

This patch also:
- adds "FORCE" to the project_config.cmake generated by
  the tfm_kconfig.py. It was not added because the Kconfig system was
  not aware of the platform settings.
- disables undefined symbol warnings because confile files to be load
  could contain non-Kconfig items, for example CMake built-in variables

Change-Id: I8fdb63bf62dd71196b37ce581dc269203d30319b
Signed-off-by: Kevin Peng <kevin.peng@arm.com>