SPM: Enhance the veneer APIs implementation for Armv8.0-M
The veneer APIs of different call types for Armv8.0-M can call
the corresponding tfm_psa_xxx directly. So that implementation
of PSA APIs for in this case is unified with the implementation
of the PSA APIs for secure.

This patch is also a preparation for performance optimization in
isolation level 2 and 3.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: If71b7bac359d65df7b83d3b376302e5467173f12