cmake: Workaround issue in fetch_remote_library and cmake 3.25.x
There appears to be an issue with cmake 3.25 and the workaround in
fetch_remote_library for GIT_SUBMODULE being an empty string.
Because of the empty strings placed by the GIT_SUBMODULE workaround the
commit hash ends up as the string "<commit>;;".
Fix this by always having another argument behind the GIT_TAG argument.

Change-Id: Ia19433d15001c270341e2a12c9845448027d5719
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>