CC312: Fix the logging macro for TF-M based deployments
The logging functions currently just rely on printf and
vprintf. But that does not work in higher level of isolations
because to reference the vprintf provided by TF-M, the function
tfm_sp_log_vprintf should be used. But the logging macro needs
to work both for runtime and bootloading binaries, hence rename
tfm_sp_log_vprintf to vprintf in able to link against that in
case it's available, or fallback to vprintf otherwise (i.e.
follow the same approach of printf). At the same time, add
automatically the LF character to the string passed in the logging
macro so that it does not need to be used in the macro instantiation.
Remove the CR as that is left to terminal interpret to add.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: Ifa18681eec8caeba59d44acf4efa48add8645ef2