docs: integration: Improve FP arch documentation
Improve the documentation of FP arch for the platforms.
The documentation was saying this should be set for the platform, but
not in which way. For most platforms this must now be set
as an additional argument to the build as most platform have not
added this configuration.
Based on the existing mps2/an521 and mps3/an552 boards setting this
variable in the preload.cmake is the right way to set it.

CONFIG_TFM_FP_ARCH does not have a default value.

Change-Id: If229f9e7f9b4875869129e991d19830ad746b6cc
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>