Build: Align PSA API test compile definitions
There are two compile definitions for PSA API IPC test
- TEST_PSA_API_IPC
- PSA_API_TEST_IPC

The first one is newly added, probably to align with the Kconfig
option.
Whilst they are two different things.
The old compile definition (the second one) should be used to avoid
duplication.

This patch addresses this issue.
Meanwhile align the crypto one with IPC as well.

Change-Id: I10a09f8a49a725de086f73c5f765a54152e4bb89
Signed-off-by: Kevin Peng <kevin.peng@arm.com>