SPM: Add memory barrier instructions
Case 1:
Add an ISB instruction after 'cpsie i' to ensure that a
pending interrupt request can be recognized immediately
after being enabled in the NVIC.

Case 2:
Add an ISB instruction after 'cpsid i' to ensure an
interrupt will not be triggered after disabling it
in the NVIC.

Case 3:
Use the DSB instruction before executing the WFI or
WFE instructions in order to ensure that CPU has
finished writing buffer.

Case 4:
A DSB instruction is executed to ensure that the write
to the CPACR register is completed. The ISB instruction
is used after the DSB to ensure that new FPU settings
are applied to subsequent floating-point instructions.

Signed-off-by: Chendi Sun <chendi.sun@arm.com>
Change-Id: I07aea70f0d796e81ddca849e4f0b4e3572bde93c