RSE: Fremont: Override LCP image boot measurement
LCP image requires special handling for boot measurement as the same
image is loaded multiple times but there is only one measurement slot.

Override boot_store_measurement to allow for special handling for LCP
image. The override uses a temporary storage for the first image then
checks if there is already a stored measurement and ensures the new
measurement is the same before continuing. Only after all images are
loaded is the measurement placed in the slot.

Signed-off-by: Arnold Gabriel Benedict <arnoldgabriel.benedict@arm.com>
Signed-off-by: Joel Goddard <joel.goddard@arm.com>
Change-Id: I8bef45977048f4d18201888780b1af9d5d5f3025