SPM: Provide option to schedule when NSPE is interrupted
By default, PendSV_Handler() will skip running the scheduler if it
determines that the NSPE was interrupted by a secure interrupt. This
avoids potentially executing lower-priority SPE work when
higher-priority NSPE work is waiting. For systems where it is known that
the secure work is always higher-priority, it is useful to be able to run
the scheduler in this situation, so this patch provides a configuration
option to support doing so.

Change-Id: I596033ec54c307b225b54412f1fa3a3f5e053481
Signed-off-by: Chris Brand <chris.brand@cypress.com>