SPM: Merge 'ffm' and 'cmsis_psa' sources
The library model referenced CMSIS sources hence its adaptation
layer name was 'cmsis_func' - and then this style was applied to
the IPC model and brought in an IPC adaptation layer with the name
'cmsis_psa'. And folder 'ffm' contains the common logic.

Now the library model is deprecated so 'cmsis_psa' is the only
adaptation layer left. As far of now, there are no obvious
requirements about referencing the common logic or the adaptation
layer individually, so let's merge common logic and the adaptation
sources into one folder 'core', and keep the folders containing
extended functionalities (Non-secure clients identification e.g.)
at the same level as 'core' with a suffix 'ext' in the folder name.

Signed-off-by: Ken Liu <Ken.Liu@arm.com>
Change-Id: Ibfd753879c60aff9adbf400b9f35c9c66de86eb9
Co-authored-by: Xinyu Zhang <xinyu.zhang@arm.com>