SPM: Allocate handles locally if no connection-based services in SFN
It is a typical function-call-like procedure when TF-M is under SFN
model and all services are stateless. Handle does not need to be
retained all the time because its lifecycle ends when the service
routine returned. Allocate handles locally in the stack so it can be
used before service routine and freed by stack-popping after its
lifecycle ends.

This patch makes the following changes:
1. Move dynamic connection pool management codes to spm_connection_pool.c.
2. Implement local static handle management in spm_local_connection.c.
3. Update architecture codes to surport operation on PSPLIM which is
   required by local static handle.

Signed-off-by: Xinyu Zhang <xinyu.zhang@arm.com>
Change-Id: I80ac0921f410ebfcd5789a970691bffc1ad5054b