ITS: Enhance the tfm_its_set/get() APIs
The tfm_its_set/get() APIs is designed to be capable of
handling assets without object size limitation. And the
object set/get logic handling is not expected to be located
in the tfm_its_req_mngr.c which handles the interface with
the spm.

This patch also saves more than 60 bytes in Flash.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: I6161df16dc4ee3324e729691118a407270ad5cd1